implements auction house => act as a database just to give all auctions to customer, when he has all the componentid of them, he can interact directly with them
