+++ 2-oracles/auction_house/scrypto/manifests/create_auction_house.manifest
@@ -1,0 +1,1 @@
+CALL_FUNCTION PackageAddress("0183625742edd40f2b8b08dac0f685b8178ce80df3c182c69bdc68") "AuctionHouse" "new";
