+++ 2-oracles/auction_house/scrypto/manifests/get_all_auctions.manifest
@@ -1,0 +1,1 @@
+CALL_METHOD ComponentAddress("026cabe9610dd13bd67b7aa0963be88e515ee347ff16b5914d9506") "get_all_auctions";
